LIU-508: Submit graph from remote git repository
This sets up the ability to submit a remotely stored graph from the remote-submit CLI. To keep with the current implementation, I have added CLI
arguments for remote files with the minimum necessary options for

    - Github or Gitlab
    - User or organisation
    - Repository name
    - Branch
    - Path to graph in the repository

This is implemented using the GitHub and GitLab APIs for GETting raw files.

By keeping with the current approach of CLI-first, these additions mean an even more convoluted command for submitting a graph. To simplify this, I've
modified our .ini configuration parsing to act as a direct drop-in replacement for the CLI. This makes it a lot easier to setup complex configurations
using remote graphs.
