+++ daliuge-engine/dlg/deploy/configs/default.ini
@@ -1,5 +1,29 @@
 ; This will be replaced by the above definition when processed
-[ENVIRONMENT]
+
+[DEPLOYMENT]
+remote=False
+submit=False
+
+[ENGINE]
+NUM_NODES = 1
+NUM_ISLANDS = 1
+ALL_NICS =
+
+[GRAPH]
+; Local
+LOGICAL_GRAPH = /tmp/HelloWorld-Universe.graph
+PHYSICAL_GRAPH =
+
+; Remote
+GITHUB = True
+; GITLAB = True
+USER_ORG = ICRAR
+REPO = EAGLE-graph-repo
+BRANCH = master
+PATH = examples/HelloWorld-Universe.graph
+
+
+[FACILITY]
 USER = 
 ACCOUNT =   
 LOGIN_NODE =
