+++ daliuge-engine/dlg/deploy/configs/setonix.ini
@@ -1,6 +1,8 @@
 ; This will be replaced by the above definition when processed
+
+
 [ENVIRONMENT]
-USER = rbunney
+USER =
 ACCOUNT = pawsey0411
 LOGIN_NODE = setonix.pawsey.org.au
 HOME_DIR = /scratch/${ACCOUNT}
