+++ daliuge-engine/test/deploy/setonix.ini
@@ -1,4 +1,4 @@
- [ENVIRONMENT]
+[ENVIRONMENT]
 ACCOUNT = pawsey0411
 USER = test
 LOGIN_NODE = setonix.pawsey.org.au
